Include twilio booking number in confirmation
Includes the booking location specific contact number when emailing the
customer with their booking request confirmation.

I moved this into a background job that runs inline for the time being.
The thinking here was that both the locations API call and email
delivery are better done out-of-band. It felt wrong to make the
locations API call in the mailer itself, too.