Wire up customer notification email
Send the customer a notification email when the booking request is
fufilled by the booking manager.

I reworked the `LocationAwareBookingRequest` into something more
generic that was suitable to any entity / model that exposes a
`location_id` (among other things).

The booking locations gem was bumped to take advantage of the newly
added helpers but primarily for the GlobalID support, which enables us
to pass these around into background jobs i.e. the mailers.