Corrects stats for fulfilment time
This should look to the difference between the `created_at` timestamps
of both the appointment and underlying booking request.

Since this calculation can be performed when the appointment is
initially persisted - and since Rails (rightly) doesn't default the
`created_at` timestamp we depend on for our calculation - I provided a
default value that serves this purpose. Rails will do the Right Thing
as far as setting this timestamp is concerned, after it is persisted.

Since the calculations are performed against seconds after the epoch
the potential for drift isn't something we should be concerned about.