Provide simple webhook support for statistics
This simply calculates the averages for fulfilment timings and the
window of time between the appointment date and the customer's first
selected slot -- per booking location.

This is then sent POSTed to the configured `STATISTICS_WEB_HOOK_URI` as
a JSON payload. Upon receipt of a non-successful status code it throws
an error. This would be handled during retries, since this would be
scheduled to execute in the context of a sidekiq worker.

Timeouts for the hook can be configured using
`STATISTICS_WEB_HOOK_TIMEOUT` and `STATISTICS_WEB_HOOK_OPEN_TIMEOUT`.