Show appointment edits as activities
Converts and displays appointment edits (provided by `audited`) as
activities in the feed.

I went with an STI approach to make the addition of further activity
types easier in future. Having the activity `message` as simple text
affords us lots of possibilities / control over how and what we display
as the activity summary.

The activities themselves are linked to the
booking request and an owning user. For those activities that do not
have a natural owner, we display 'Somebody' as a placeholder.