Poll for new activities
Polls for new activities based on the interval configured in
`POLL_INTERVAL_MILLISECONDS`.

When new activites are added they are animated in. If multiple
activites were added, they are all rendered as you'd expect.

In my development environment, `304` calls return in the order of ~8ms.
Calls that yield rendered activities return in the order of ~20ms. This
short-polling approach will clearly not scale beyond ~100 users but if
that becomes an issue we can offload to pusher or another, similar
service.