Add message activities
Gives the booking manager the ability to leaves messages against a
booking request and / or an appointment.

There is no validation and the booking manager can leave a blank
message if they like.

Polling would cause message-based activities to appear twice for the
user creating them, since they'd appear from the parsed response after
AJAX submission of the message form and from the polled response. To
workaround this I perform a simple check for existence of that
particular row - tagged with a known ID.

This approach is not particularly efficient, but is simple enough and
it works well in practice.