Handle incoming drop webhooks from MailGun
When MailGun sends an inbound drop webhook we look up the latest
matching booking request from the original recipient's email and create
an activity containing the actual reasoning for the dropped message.

Any problems raise an error that bubbles up to bug snag.

Requires the `MAILGUN_API_TOKEN` to be set.