Ensure inbound hooks are filtered
Changes our approach to further validate the required properties of a
message. Essentially we only care about online booking messages from
production that have an associated recipient.

There are cases when the booking request may not be found that we
cannot action. Such as during a race when the initial email was sent,
but subsequently the booking manager changed the client email during
fulfillment so we cover this fully.

Due to the way webhooks are configured we cannot route inbound requests
on to the correct environment without our own custom routing
infrastructure. In which case we'll silently drop messages meant for
other environments outside of production.

Ideally we'd have multiple mailgun configurations to match per project
but that's perhaps a step for the future.

We also drop messages that did not actually originate from online
bookings / planner.