Allow administrators to switch locations
This allows users with the necessary privileges to switch their booking
location 'ownership' thus allowing them to see bookings / appointments
from any location.

Prior to this the user would have to go to signon and change their
organisation then log out / in.

Note: The chosen organisation will be reverted upon the next sign in,
depending on what organisation is assigned in the signon application.