Simplify drop messages
The drop messages sometimes reflect the SMTP language directly,
depending on the recipient address. This is probably too complex for
the bookings managers to understand and cannot be actioned by them.

We'll continue to store the original SMTP language and reasoning, but
only display a friendly message to the user that they can action.