Clarify intent of DC pot flag
Since we already permit `true` or `false` (despite `false` never having
been provided as the client does not permit this) and given we need to
report this to the booking managers - we now accept `false` to signify
the customer was 'not sure' if they had a DC pot and were thus eligible.

This will require some rework to ensure the client(s) pass `false` in
the case of the 'not sure' answer being provided.

All existing records are marked 'true' and cannot be affected
historically. We can safely assume these bookings are DC pot confirmed.