Scheduled customer appointment reminders
We will now send scheduled appointment reminders at 3PM every day. This
will send an email reminder to the customer, containing their
appointment details. We are hoping this causes an uplift in appointment
attendance and in turn reduces the no-show rate.

The system guards against multiple reminder deliveries based on the
presence of the newly introduced `ReminderActivity` which also shows in
the activity feed as a record of this reminder being sent.

For this to run we will need to schedule creation of the
`AppointmentRemindersJob`, that in turn fans out a job per appointment
that needs a reminder sent.