Allow modifications to booking when fulfilling
We now allow the booking managers to modify the booking request in
place before transitioning to the appointment. The booking request
itself is still considered immutable, so this required a 1:1 mirroring
of booking request attributes to appointment attributes.

This also required some duplication of validation rules between the
resultant appointment and the form object. Once we add the further,
editable attributes on the appointment edit screen I can consolidate
these rules into once place.