Don't audit when copying attributes
This makes no sense to log and shouldn't be pointed out to the customer
by way of an appointment or booking request activity entry.

Amendments to migrations are usually a bad idea but we're fine here
since this has yet to run and in those environments where it has it can
be deeemed idempotent.