Reassign requests by booking location
When a booking location is assigned to another, the logic around
reassignment is slightly different. First we need to check if the given
location ID refers to a booking location - if it does we can assume the
booking location is being reassigned - so we only update bookings
matching their booking location ID with the given location ID.

Otherwise we fallback to the previous logic and update the booking
location ID of bookings matching the given location ID. This can be
confusing since bookings can take place at their booking location - in
which case the booking location ID and location ID are the same.