Don't display schedules for hidden locations
Booking managers seem to be confused about how to treat hidden
locations so some have schedules and some don't. This isn't necessary
since they don't appear to customers during the booking journey and
therefore cannot be booked.

Some locations have a great number of hidden locations and picking
through them when managing schedules is a pain, so we elected to not
display them in the schedules grid.