Validate time presence
Since we removed the preselection of time during the fulfilment of
bookings, we've seen several instances where agents or booking managers
forget to select the appointment time. This causes errors further down
in the mapper since their is an attempt to format the time.

Adding this validation will stop that from occuring and inform the
person attempting to place the booking exactly where they have gone
wrong.