Remove postal confirmation banner message
When the customer has requested a postal confirmation we already display
the link to 'Postal address' in the left side bar which allows them to
modify the customer supplied address. Since we now automatically process
printed confirmations via notify we no longer need to display the banner
as a cue for the agent to manually print the letter.