Don't permit a continual cycle of drop notifications
When a booking manager creates an appointment and the confirmation email
to themselves (in the role of customer) fails, they will be sent the
booking manager's regular email failure notification email. If this
subsequently fails, this enters an infinite loop of bounce/retry until
the application queue - in this instance, redis - falls over with
out-of-memory exceptions.

This change ensures that whenever the booking's customer email matches
any of the found booking manager emails the drop processing is aborted
and a successful status is reported to mailgun, ensuring they do not
continually retry.