Remove old-style slots
Addresses the issues with slots and formatting. Prior to this we were
using combination of date/start/end for bookable slots and this is
complicated for many reasons. Moving towards a simple date/time for
`start_at` and `end_at` removes most of the complexity.

This was also a good time to remove redundant old-style slot features.