Correctly track audits/changes
We started to use nested forms to alter the booking request attributes
but have realised that any attributes altered on the booking request do
not trigger audits on the owning parent appointment record. This is
problematic for many reasons but namely that we have no audit log of
changes an agent makes to the third party or BSL attributes. Also,
changes to these attributes were not triggering change notification
emails for booking managers - so potentially they have been missing out
on critical notifications.

The change was non-trivial and required some combining of audited
attributes to track the changes manually, though in concert with the
audited gem's behaviour around associated audits.

This seems to work relatively well for now.